Extract .bloomSource import to its own file and make its API stateless (BL-16502)
Two refactors of the .bloomSource import feature, no behavior change:

1. Factor the import code (~740 lines) out of the already-bloated
   CollectionModel.cs into a new partial-class file,
   CollectionModel.BloomSourceImport.cs: the file picker, the duplicate
   pre-flight check, and the import itself (edit / derivative / replace /
   add-copy). Saving/exporting a book AS a .bloomSource stays in
   CollectionModel.cs. Also drops two now-unused usings there.

2. Remove the server-side state the feature carried between three API calls.
   Previously the chosen file paths lived in a _bloomSourceFilesToImport field
   across chooseBloomSourceFilesToImport -> bloomSourceImportDuplicateInfo ->
   importBloomSource. Now the backend is stateless commands that the front-end
   drives:
   - ChooseBloomSourceFilesToImport() returns the chosen string[] paths.
   - The choose endpoint returns { files, anyDuplicates, canReplace } in one
     reply, so the separate bloomSourceImportDuplicateInfo endpoint is gone.
   - importBloomSource reads the paths from the POST body; the front-end holds
     them in React state and passes them back.

The file-system mechanics (zip, validation, folder moves, Team Collection
handling, rollback) stay in C# where they belong; only the orchestration state
moved to the front-end.
